requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
celery==5.3.6
redis==5.0.1
rapidfuzz==3.6.1
python-dateutil==2.8.2
alembic==1.13.1
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
python-multipart==0.0.6
phonetics==1.0.5
requests==2.31.0
websockets==12.0
openpyxl==3.1.2
pandas==2.1.3
openai==1.3.0
